backend/requirements.txt
# kwebbie - Copyright 2025 milbert.ai
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0
python-multipart==0.0.9

# Database - SQLite and PostgreSQL with async support
sqlalchemy==2.0.25
aiosqlite==0.19.0
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyotp==2.9.0

# Validation and serialization
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# WebSocket
python-socketio==5.11.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.4

# XML parsing (for tool output)
lxml==5.1.0
xmltodict==0.13.0

# Report generation
weasyprint==60.2
python-docx==1.1.0
jinja2==3.1.4
markdown==3.5.2

# Encryption
cryptography==42.0.5

# Task queue
celery[redis]==5.3.6
redis==5.0.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
orjson==3.9.13
aiofiles==23.2.1

# Optional integrations
slack-sdk==3.27.0
atlassian-python-api==3.41.10

# Development/Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
